Restructure and organize folder structure
