+++ frontend/src/components/Tasks/ListFooter.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import '../styles/CSS/ListContainer.css'
+import '../../styles/CSS/ListContainer.css'
 
 const ListFooter = () => {
   return (
