+++ frontend/src/pages/Tasks.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import ListContainer from "../components/ListContainer";
+import ListContainer from "../components/Tasks/ListContainer";
 import HomeButton from '../components/buttons/HomeButton';
 
 const Tasks = () => {
